awsCdk/awsCdkProject/pyproject.toml
[tool.poetry]
name = "awsInfrastructure"
version = "0.1.0"
description = ""
authors = ["Michael C Hoffman <michaelchoffman@gmail.com>"]
packages = [
    { include = "awsInfrastructure" }
]

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = { version = "*", allow-prereleases = true}
"aws-cdk.aws-lambda-python-alpha" = "^2.25.0-alpha.0"

[tool.poetry.dev-dependencies]
safety = { version = "*" }
black = { version = "*", allow-prereleases = true }
flake8 = { version = "*" }
isort = { version = "*" }
mypy = { version = "*" }
pytest-aio = { version = "*" }
pytest-assume = { version = "*" }
pytest-asyncio = { version = "*" }
pytest-black = { version = "*" }
pytest-colordots = { version = "*" }
pytest-cov = { version = "*" }
pytest-datadir = { version = "*" }
pytest-depends = { version = "*" }
pytest-flake8 = { version = "*" }
pytest-html = { version = "*" }
pytest-metadata = { version = "*" }
pytest-mock = { version = "*" }
pytest-pylint = { version = "*" }
pytest-sugar = { version = "*" }
pytest-isort = { version = "*" }
pytest-mypy = { version = "*" }
pytest = { version = "*" }
pytest-xdist = { version = "*" }

[tool.pytest.scripts]
mypyTests = "scripts.codeFormat:mypyTests"
isortTests = "scripts.codeFormat:isortTests"
isortFix = "scripts.codeFormat:isortFix"
blackTests = "scripts.codeFormat:blackTests"
blackFix = "scripts.codeFormat:blackFix"
pylintTests = "scripts.codeFormat:pylintTests"
flake8Tests = "scripts.codeFormat:flake8Tests"
pytestCovTerm = "scripts.codeFormat:pytestCovTerm"
pytestTestReports = "scripts.codeFormat:pytestTestReports"
pytestCovReports = "scripts.codeFormat:pytestCovReports"

[tool.pytest.ini_options]
python_classes = "*Test"
python_functions = "*Test"
python_files = "*Tests.py"
addopts = "-v"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(levelname)s %(filename)s %(funcName)s %(lineno)d %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
testpaths = [
    "awsInfrastructure",
    "tests",
]
markers = [
    "unitTests: unit tests",
    "integrationTests: integration tests",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.coverage.run]
source = ["awsInfrastructure"]
omit = [
    "awsInfrastructure/__main__.py",
    "tests/conftest.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    'if self\.debug',
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    'if __name__ == .__main__.:',
]
